+++ .gitignore
@@ -1,0 +1,2 @@
+Backend/node_modules/
+Backend/.env
